DB: define field types outside of the models
