+++ beets/dbcore/types.py
@@ -14,13 +14,26 @@
 
 """Representation of type information for DBCore model fields.
 """
+
+import re
+import sys
+import time
 import typing
 from abc import ABC
 from typing import Any, Generic, List, TypeVar, Union, cast
 
-from beets.util import str2bool
+from beets import config, util
 
-from .query import BooleanQuery, FieldQuery, NumericQuery, SubstringQuery
+from .query import (
+    BLOB_TYPE,
+    BooleanQuery,
+    DateQuery,
+    DurationQuery,
+    FieldQuery,
+    NumericQuery,
+    PathQuery,
+    SubstringQuery,
+)
 
 
 class ModelType(typing.Protocol):
@@ -310,7 +323,138 @@ class Boolean(Type):
         return str(bool(value))
 
     def parse(self, string: str) -> bool:
-        return str2bool(string)
+        return util.str2bool(string)
+
+
+class PathType(Type):
+    """A dbcore type for filesystem paths.
+
+    These are represented as `bytes` objects, in keeping with
+    the Unix filesystem abstraction.
+    """
+
+    sql = "BLOB"
+    query = PathQuery
+    model_type = bytes
+
+    def __init__(self, nullable=False):
+        """Create a path type object.
+
+        `nullable` controls whether the type may be missing, i.e., None.
+        """
+        self.nullable = nullable
+
+    @property
+    def null(self):
+        if self.nullable:
+            return None
+        else:
+            return b""
+
+    def format(self, value):
+        return util.displayable_path(value)
+
+    def parse(self, string):
+        return util.normpath(util.bytestring_path(string))
+
+    def normalize(self, value):
+        if isinstance(value, str):
+            # Paths stored internally as encoded bytes.
+            return util.bytestring_path(value)
+
+        elif isinstance(value, BLOB_TYPE):
+            # We unwrap buffers to bytes.
+            return bytes(value)
+
+        else:
+            return value
+
+    def from_sql(self, sql_value):
+        return self.normalize(sql_value)
+
+    def to_sql(self, value):
+        if isinstance(value, bytes):
+            value = BLOB_TYPE(value)
+        return value
+
+
+class MusicalKey(String):
+    """String representing the musical key of a song.
+
+    The standard format is C, Cm, C#, C#m, etc.
+    """
+
+    ENHARMONIC = {
+        r"db": "c#",
+        r"eb": "d#",
+        r"gb": "f#",
+        r"ab": "g#",
+        r"bb": "a#",
+    }
+
+    null = None
+
+    def parse(self, key):
+        key = key.lower()
+        for flat, sharp in self.ENHARMONIC.items():
+            key = re.sub(flat, sharp, key)
+        key = re.sub(r"[\W\s]+minor", "m", key)
+        key = re.sub(r"[\W\s]+major", "", key)
+        return key.capitalize()
+
+    def normalize(self, key):
+        if key is None:
+            return None
+        else:
+            return self.parse(key)
+
+
+class DurationType(Float):
+    """Human-friendly (M:SS) representation of a time interval."""
+
+    query = DurationQuery
+
+    def format(self, value):
+        if not config["format_raw_length"].get(bool):
+            interval = int(value or 0.0)
+            return "%i:%02i" % (interval // 60, interval % 60)
+        else:
+            return value
+
+    def parse(self, string):
+        try:
+            # Try to format back hh:ss to seconds.
+            return util.raw_seconds_short(string)
+        except ValueError:
+            # Fall back to a plain float.
+            try:
+                return float(string)
+            except ValueError:
+                return self.null
+
+
+class DateType(Float):
+    # TODO representation should be `datetime` object
+    # TODO distinguish between date and time types
+    query = DateQuery
+
+    def format(self, value):
+        return time.strftime(
+            config["time_format"].as_str(), time.localtime(value or 0)
+        )
+
+    def parse(self, string):
+        try:
+            # Try a formatted date string.
+            return time.mktime(
+                time.strptime(string, config["time_format"].as_str())
+            )
+        except ValueError:
+            # Fall back to a plain timestamp number.
+            try:
+                return float(string)
+            except ValueError:
+                return self.null
 
 
 # Shared instances of common types.
