+++ beetsplug/metasync/amarok.py
@@ -22,7 +22,6 @@ from time import mktime
 from xml.sax.saxutils import quoteattr
 
 from beets.dbcore import types
-from beets.library import DateType
 from beets.util import displayable_path
 from beetsplug.metasync import MetaSource
 
@@ -43,8 +42,8 @@ class Amarok(MetaSource):
         "amarok_score": types.FLOAT,
         "amarok_uid": types.STRING,
         "amarok_playcount": types.INTEGER,
-        "amarok_firstplayed": DateType(),
-        "amarok_lastplayed": DateType(),
+        "amarok_firstplayed": types.DateType(),
+        "amarok_lastplayed": types.DateType(),
     }
 
     query_xml = '<query version="1.0"> \
