+++ beetsplug/playlist.py
@@ -18,8 +18,7 @@ import tempfile
 from typing import Sequence
 
 import beets
-from beets.dbcore.query import InQuery
-from beets.library import BLOB_TYPE
+from beets.dbcore.query import BLOB_TYPE, InQuery
 from beets.util import path_as_posix
