+++ beetsplug/types.py
@@ -15,7 +15,6 @@
 
 from confuse import ConfigValueError
 
-from beets import library
 from beets.dbcore import types
 from beets.plugins import BeetsPlugin
 
@@ -42,7 +41,7 @@ class TypesPlugin(BeetsPlugin):
             elif value.get() == "bool":
                 mytypes[key] = types.BOOLEAN
             elif value.get() == "date":
-                mytypes[key] = library.DateType()
+                mytypes[key] = types.DateType()
             else:
                 raise ConfigValueError(
                     "unknown type '{}' for the '{}' field".format(value, key)
