+++ test/test_library.py
@@ -30,6 +30,7 @@ from mediafile import MediaFile, UnreadableFileError
 import beets.dbcore.query
 import beets.library
 from beets import config, plugins, util
+from beets.dbcore.types import DateType, DurationType, MusicalKey, PathType
 from beets.test import _common
 from beets.test._common import item
 from beets.test.helper import BeetsTestCase, ItemInDBTestCase
@@ -1341,7 +1342,7 @@ class LibraryFieldTypesTest(unittest.TestCase):
     """Test format() and parse() for library-specific field types"""
 
     def test_datetype(self):
-        t = beets.library.DateType()
+        t = DateType()
 
         # format
         time_format = beets.config["time_format"].as_str()
@@ -1354,7 +1355,7 @@ class LibraryFieldTypesTest(unittest.TestCase):
         assert t.null == t.parse("1973-11-29")
 
     def test_pathtype(self):
-        t = beets.library.PathType()
+        t = PathType()
 
         # format
         assert "/tmp" == t.format("/tmp")
@@ -1364,7 +1365,7 @@ class LibraryFieldTypesTest(unittest.TestCase):
         assert np(b"/tmp/\xc3\xa4lbum") == t.parse("/tmp/\u00e4lbum/")
 
     def test_musicalkey(self):
-        t = beets.library.MusicalKey()
+        t = MusicalKey()
 
         # parse
         assert "C#m" == t.parse("c#m")
@@ -1372,7 +1373,7 @@ class LibraryFieldTypesTest(unittest.TestCase):
         assert "Not c#m" == t.parse("not C#m")
 
     def test_durationtype(self):
-        t = beets.library.DurationType()
+        t = DurationType()
 
         # format
         assert "1:01" == t.format(61.23)
